Make VerificationCache header-only to fix Workspaces link (LNK2001)
The Workspaces duplicate transport (WorkspacesLib) compiles the shared
two_way_pipe_message_ipc_impl.h, so it embeds a VerificationCache member but does not link the auth
translation unit. The pimpl VerificationCache's out-of-line ctor/dtor were therefore unresolved when
linking PowerToys.WorkspacesLauncher.exe / PowerToys.WorkspacesWindowArranger.exe.

Make VerificationCache fully header-only/inline (like CallerPolicy) so it introduces no out-of-line
symbols; any TU that includes the header links with no extra dependency. Verified WorkspacesLauncher
and WorkspacesWindowArranger (Release) now link, interop Release + runner Debug build clean, and the
7 gate tests pass.

Copilot-Session: 1d8f85ec-aaee-468a-9348-ba79b059cbaf
